feat: create a binary_helper
Separate the functions that handle that checks if binary is
1111, 1111.1, -1111, or -1111.1.

Fix all the imports problem.
