+++ src/menu/binary/__tests__/test_binary_to_decimal.c
@@ -1,4 +1,5 @@
 #include "../../../utils/utils.h"
+#include "../binary_helper.c"
 #include "../binary_to_decimal.c"
 #include <assert.h>
 
@@ -14,7 +15,8 @@ void test_binary_to_decimal(void) {
   text_green("is_positive -> should not accept positive binary with dot", 10);
   add_new_line(1);
   assert(is_positive("-1111") == 0);
-  text_green("is_positive -> should not accept negative binary without dot", 10);
+  text_green("is_positive -> should not accept negative binary without dot",
+             10);
   add_new_line(1);
   assert(is_positive("-1111.1") == 0);
   text_green("is_positive -> should not accept negative binary with dot", 10);
@@ -22,16 +24,22 @@ void test_binary_to_decimal(void) {
 
   // 1111.1 values.
   assert(is_positive_with_dot("1111.1") == 1);
-  text_green("is_positive_with_dot -> should accept positive binary with dot", 10);
+  text_green("is_positive_with_dot -> should accept positive binary with dot",
+             10);
   add_new_line(1);
   assert(is_positive_with_dot("1111") == 0);
-  text_green("is_positive_with_dot -> should not accept positive binary without dot", 10);
+  text_green(
+      "is_positive_with_dot -> should not accept positive binary without dot",
+      10);
   add_new_line(1);
   assert(is_positive_with_dot("-1111") == 0);
-  text_green("is_positive_with_dot -> should not accept negative binary without dot", 10);
+  text_green(
+      "is_positive_with_dot -> should not accept negative binary without dot",
+      10);
   add_new_line(1);
   assert(is_positive_with_dot("-1111.1") == 0);
-  text_green("is_positive_with_dot -> should not accept negative binary with dot", 10);
+  text_green(
+      "is_positive_with_dot -> should not accept negative binary with dot", 10);
   add_new_line(1);
 
   // -1111 values.
@@ -42,7 +50,8 @@ void test_binary_to_decimal(void) {
   text_green("is_negative -> should not accept negative binary with dot.", 10);
   add_new_line(1);
   assert(is_negative("1111") == 0);
-  text_green("is_negative -> should not accept positive binary without dot.", 10);
+  text_green("is_negative -> should not accept positive binary without dot.",
+             10);
   add_new_line(1);
   assert(is_negative("1111.1") == 0);
   text_green("is_negative -> should not accept positive binary with dot.", 10);
@@ -50,16 +59,22 @@ void test_binary_to_decimal(void) {
 
   // -1111.1 values.
   assert(is_negative_with_dot("-1111.1") == 1);
-  text_green("is_negative_with_dot -> should accept negative binary with dot", 10);
+  text_green("is_negative_with_dot -> should accept negative binary with dot",
+             10);
   add_new_line(1);
   assert(is_negative_with_dot("-1111") == 0);
-  text_green("is_negative_with_dot -> should not accept negative binary without dot", 10);
+  text_green(
+      "is_negative_with_dot -> should not accept negative binary without dot",
+      10);
   add_new_line(1);
   assert(is_negative_with_dot("1111") == 0);
-  text_green("is_negative_with_dot -> should not accept positive binary with dot", 10);
+  text_green(
+      "is_negative_with_dot -> should not accept positive binary with dot", 10);
   add_new_line(1);
   assert(is_negative_with_dot("1111.1") == 0);
-  text_green("is_negative_with_dot -> should not accept positive binary without dot", 10);
+  text_green(
+      "is_negative_with_dot -> should not accept positive binary without dot",
+      10);
   add_new_line(1);
 
   // Compute decimal 1111 value.
