+++ src/menu/binary/__tests__/test_binary_validate.c
@@ -13,8 +13,11 @@ void test_binary_validate(void) {
   text_green("binary_validate -> should accept negative binary values", 10);
   add_new_line(1);
   assert(binary_validate("0001.0001") == 1);
-  text_green("binary_validate -> should accept binary values with binary dot", 10);
+  text_green("binary_validate -> should accept binary values with binary dot",
+             10);
   add_new_line(1);
   assert(binary_validate("-0001.0001") == 1);
-  text_green("binary_validate -> should accept negative binary values with binary dot", 10);
+  text_green(
+      "binary_validate -> should accept negative binary values with binary dot",
+      10);
 }
